mod addN
